test: migrate raw http request test
